Split up arrow_cast::list
